moved tests from mailroom4.py to test_mailroom4.py.  Remove donor parameter from mailroom4 functions
